+++ NEWS.md
@@ -5,9 +5,10 @@ about regenerating parameters from the synthetic exposures and compare with that
 from real exposures if `verbose > 0`.
 
 ### Added
-* Added new argument `tumor.marker.name` in function `GenerateSyntheticTumors`. 
+* Added new argument `tumor.marker.name` in function `GenerateSyntheticTumors`.
 
-* Added new argument `sig.params` in exported function `GetSynSigParamsFromExposures`.
+* Added new argument `sig.params` in exported functions
+`GetSynSigParamsFromExposures` and `GenerateSyntheticTumors`.
 
 <br>